radiomanager-frontend: Improve api module structure
